# internal/dto/user/update_user.go
package user

type FieldUpdateDetails struct {
	Email       string `json:"email"`
	PhoneNumber string `json:"phoneNumber"`
	FirstName   string `json:"firstName"`
	LastName    string `json:"lastName"`
}